requirements.txt
flask
flask-compress
gunicorn
requests
joblib
python-kraken-sdk
xgboost
scikit-learn
web3==6.15.0
sentry-sdk[flask]
certifi
python-dotenv
pandas
